Converted package bcema version BLP_LIB_BDE_BCE_2.5.0 from clearcase

+++ groups/bce/bcema/bcema_fixedpool.h
@@ -427,7 +427,7 @@ int bcema_FixedPool::objectSize() const
 inline
 int bcema_FixedPool::poolSize() const
 {
-    return d_nodes.size();
+    return static_cast<int>(d_nodes.size());
 }
 
 }  // close namespace BloombergLP
